Revert "Feature/multi device pedometer"
